Improves FINALLY sample handling
Adds handling for the FINALLY samples to indicate when a sample is unavailable.

The "FINALLY" samples are only available for a subset of the
samples. This change ensures that when a "FINALLY" sample is not
available, a message is displayed to the user indicating that the
sample is not available and why.